personalization-service/requirements.txt
python-dotenv
langchain
langchain-aws
boto3
flask
flask-cors
boto3
sqlalchemy
psycopg2-binary
pgvector
requests
langchain_community
langgraph
